convert the app logic and UI to deal with stage 2 scenario
This change is extremely invasive because it touches pretty much every aspect
of the existing code. With the exception of code dealing with the hardware
connected to a feeder, everything else needs to be changed to now work on lists
and keep the micro feeder id in mind.

I assume that this code is still buggy, goal number one is to ensure that it
works exactly the same as the existing code in the case of an integrated feeder
that both has local SDR(s) and talks to aggregators.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>